requirements.txt
requests~=2.31.0
python-dotenv~=1.1.1
matplotlib~=3.10.3
seaborn~=0.13.2
pandas~=2.1.4
streamlit~=1.46.1
pycountry~=24.6.1
plotly~=6.2.0